LayoutTests/fast/dom/shadow/shadow-hierarchy-exception.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<div id=container>
  <div>
      <span></span>
  </div>
</div>
<script>
description('Appending an ancestor to a shadow tree should throw an exception');
var container = document.getElementById('container');
var span = container.querySelector('span');
var shadow = span.webkitCreateShadowRoot();
var shadowDiv = shadow.appendChild(document.createElement('div'));
shouldThrow('shadowDiv.appendChild(container)');
</script>
<script src="../../js/resources/js-test-post.js"></script>
